EntityID is now accessible through a Box2D body
Bodys' user data is a pointer to the PhysicsComponent, which now
holds the entity ID. If any more data is required other than the
ID, another struct will be used (ie. EntityIdentifier). However
this must be dynamically allocated and will be a pain to
implement efficiently, hence why the already-managed input
component is currently used.